refactors the headcount analysis and puts methods into helper
