+++ lib/lib_helper.rb
@@ -1,0 +1,1 @@
+require 'pry'
